Language mapping, file hierarchy, emergencies
Added files en.lproj and fr.lproj to localize french and english text to the appropriate labels. An initial file hierarchy has been created, but might be subject to change to be in accordance with MVVM. The "more" section has been created, with emergencies content table view linked to it.