Duplicate models
In order to persist entities in the DB, they may need to includes additional attributes. The additional attributes would be used to retrieve the entities later on.

This is needed because most request to the webservice aims to retrieve data by one or multiple attributes. However, the response may not includes those specific attributes in the model. If we don't manually set the attributes before saving the entity in the DB, it would be very difficult to the retrieve the entities later on from the DB, for example, by using the WHERE clause.

The problem is that Moshi needs to ignore the additional attributes, which is possible by using transient. However, Room would also ignore the additional attributes.

Furthermore, even though using the same model for Moshi and Room minimizes duplication, the response model and the entity model are, as a result, tightly coupled.

A solution is to "duplicate" the models and use mappers to transform the api's models into entities for Room.