Add CoursRepository
When requesting the courses, Signets doesn't provide the grade. It only provides the letter grade. However, the UI may wish to display the grade for a given a course if it doesn't have a letter grade yet. As a result, when creating the call, CoursRepository needs to apply transformations to get the grade for each course.