Optimisation du processus d'intégration continue avec Travis CI (#94)
* Optimize Android Travis build process based on previous expriments

A Travis-CI build might take at least 25 minutes, this in itself
might not be the best solution available. Because of that, let's
optimize the build flow in an attempt to reduce build times

- Install all the required Android SDK components into before_install.sh
- Install the canary version (29.0.3) of the emulator which supports x86
images without acceleration. (Travis doesn't support nested virtualization)
- Move the emulator creation/start-up into the install section of
travis.yml

More info about the emulator optimization over here:
https://travis-ci.community/t/android-emulators-not-starting-for-the-last-few-days-late-march-2019/2871/11

* travis.yml: Set correctly the envrionment variables

Since the Travis-CI image for Android is a bit outdated, adjust
PATH so it can be used with the newer executables from the SDK.

Based on:
https://travis-ci.org/ankidroid/Anki-Android/jobs/516486600/config

* travis.yml: Use debug version when building/testing

* travis.yml: Remove sudo: false attribute

More info:
https://docs.travis-ci.com/user/reference/overview/#deprecated-virtualization-environments

* travis.yml: Remove Gradle locks before cache

* Adapt build for Ubuntu Xenial (16.04)

- Change the OS to Xenial with minimal language in travis.yml
- Download and extract the Android SDK manually
- Change the PATH accordingly in travis.yml
- Downgrade the version of JDK used to JDK8 (A more recent version is used)
- Add android-wait-for-emulator script for waiting on a booted instance
of the Android emulator for instrumentation tests

Configuration snippets based from:
https://travis-ci.com/G00fY2/android-ci-testproject/jobs/187085499/config

* Add some documentation for Travis related stuff

* Build only if the project files have been changed

Only build projects if actual code has changed. Since there are 2
projects in one repo. One project might always get built without
any seen changes (For example an iOS commit triggers an Android build)

For this reason, cancel a build only if the related commits
(Eg. From a PR) isn't affecting the other's project.

Don't apply this rule for travis.yml which might be used for testing
Travis-CI infrastructure.

* Add executable permission to Travis-CI scripts

* Use variable expansion for checking the build

* build-check.sh: Only execute "git diff" once

* travis.yml: Little clean up

* Transfer Android beta deployment process to travis.yml

Since the keys are already decrypted in the before_install step,
it is not necessary to decrypt them again.

In that case, we only have a few commands and thus, they should be executed
directly in the .yml file.

* Android: Remove unused variable on build_install.sh

* travis.yml: Rename ETS-mobile paths to ETSMobile for iOS tests

* .travis.yml: Correctly specify the OS and the language