Fix schedule order (#110)
* Fix grades courses loading forever after visiting grades details

When the grade in the details screen would be updated, the grade displayed in the courses grades screen would also be updated. If the user exited the screen while the data is being loaded, the courses grades screen could remain in a loading status.

* Fix seances order

The seances were displayed in the wrong order. Furthermore, instead of sorting them in the use cases, we now sort them by adding a ORDER BY clause in the dao.