Dashboard cards suspend function (#149)
* Set dashboardCards methods as a suspend function in DashboardCardRepository

Return the channel directly instead of using the GlobalScope to produce a channel and consuming the first channel and offering the values to the second one.

The GlobalScope is normally used to launch top-level coroutines and does not scope the work properly. We now also declare FetchDashboardCardsUseCase's invoke operator as a suspend function. This function is called within the DashboardViewModel's scope. As a result, we now scope the work to the ViewModel, so that if the ViewModel is cleared, the work is canceled automatically.

* Update some deps