Add LastValueReceiveChannel to keep last value
For now, the class is only used to keep the last value of universal code dialog visibility. On Android, when a configuration change occurs, the Android ViewModel will keeps its data. However, as we simply use Channels and converted them into LiveDatas, the last values of the Channels weren't not getting saved. Before, when we didn't use Kotlin MP, we were declaring the LiveData inside the ViewModel and last value would be kept. When the observers would be active again, the LiveData would dispatch the last value. Channels don't have this feature since they are don't hold the data when it has been consumed. The problem is noticeable especially with the dialog. When the user rotated the screen, it would not reappear again. To fix this, we add a class called LastValueReceiveChannel that will keep the last value submitted and use a Channel by delegation.