[skip travis] Add login with MonETS (#181)
* Add ktor

* Use ktor to implement HTTP client

* Add LoginRepository to shared module

* Add CheckUserCredentialsAreValidUseCase

* Add FetchSavedUserCredentialsUseCase

* Add LoginWithSavedCredentials

* Add LoginViewModel

* Move old LoginViewModel to splash directory

It's still used for the splash screen.

* Use shared LoginViewModel in LoginFragment

* Add SplashViewModel

* Add ClearUserDataUseCase to shared module

* Remove old files

* Clear shared preferences on logout

* Use TextInputLayout's endIcon properties

* Fix field error message not dismissed when no error

* Fix error message not shown

* Fix app bar layout expanded on login screen

* Add localized strings to `shared` module

Provides a temp solution to shared localized strings between the two platforms

* Add LastValueReceiveChannel to keep last value

For now, the class is only used to keep the last value of universal code dialog visibility. On Android, when a configuration change occurs, the Android ViewModel will keeps its data. However, as we simply use Channels and converted them into LiveDatas, the last values of the Channels weren't not getting saved. Before, when we didn't use Kotlin MP, we were declaring the LiveData inside the ViewModel and last value would be kept. When the observers would be active again, the LiveData would dispatch the last value. Channels don't have this feature since they are don't hold the data when it has been consumed. The problem is noticeable especially with the dialog. When the user rotated the screen, it would not reappear again. To fix this, we add a class called LastValueReceiveChannel that will keep the last value submitted and use a Channel by delegation.

* Use LastValueReceiveChannel with field error message and add some comments

Before, when the user changed the language on its device, the language of the message would not change.

* Remove Event usage in LoginFragment

Not needed since Channels don't keep the last value

* Remove Event and EventObserver class from shared module

* Rename `LoginWithSavedCredentials`to `LoginWithSavedCredentialsUseCase`