Allow user to access the app right away if he has already successfully login before
Before, `FetchSavedUserCredentialsUseCase` would submit the saved credentials to the API to check if they were valid. This prevented the user from accessing the app if the device was offline. Furthermore,
the app's startup would be slowed down. Even though the request didn't take a while to finish, it was still noticeable.